create ics on download in resource route
Resource route: event/$slug/icsDownload.tsx

Undo prisma migration:
- Removed ics column from event table
