+++ app/routes/event/$slug/icsDownload.tsx
@@ -1,0 +1,148 @@
+import type { DateArray } from "ics";
+import * as ics from "ics"; // ics is undefined
+import { LoaderFunction } from "remix";
+import { getParamValueOrThrow } from "~/lib/utils/routes";
+import { transformAbsoluteURL } from "~/lib/utils/string";
+import { MakeOptional } from "~/lib/utils/types";
+import { getEventBySlugOrThrow } from "./utils.server";
+
+type EventWithRelations = Awaited<ReturnType<typeof getEventBySlugOrThrow>>;
+
+// TODO: Add status (CONFIRMED/CANCELLED) to the ics file (see #437)
+// TODO: Add organizer to the ics file (see #432)
+// see https://www.npmjs.com/package/ics
+export function createIcsString(
+  event: MakeOptional<
+    Pick<
+      EventWithRelations,
+      | "id"
+      | "startTime"
+      | "endTime"
+      | "name"
+      | "slug"
+      | "createdAt"
+      | "updatedAt"
+      | "description"
+      | "tags"
+      | "venueCity"
+      | "venueName"
+      | "venueStreet"
+      | "venueStreetNumber"
+      | "venueZipCode"
+      | "conferenceLink"
+      | "conferenceCode"
+    >,
+    | "description"
+    | "tags"
+    | "venueCity"
+    | "venueName"
+    | "venueStreet"
+    | "venueStreetNumber"
+    | "venueZipCode"
+    | "conferenceLink"
+    | "conferenceCode"
+  >,
+  absoluteEventUrl: string
+) {
+  const location: string[] = [];
+  if (event.conferenceLink) {
+    location.push(`Konferenzlink: ${event.conferenceLink}`);
+  }
+  if (event.conferenceCode) {
+    location.push(`Zugangscode zur Konferenz: ${event.conferenceCode}`);
+  }
+  if (event.venueName) {
+    location.push(event.venueName);
+  }
+  if (event.venueStreet) {
+    const fullStreet = `${event.venueStreet} ${event.venueStreetNumber || ""}`;
+    location.push(fullStreet.trim());
+  }
+  if (event.venueZipCode) {
+    location.push(event.venueZipCode);
+  }
+  if (event.venueCity) {
+    location.push(event.venueCity);
+  }
+  let tagTitles: string[] = [];
+  if (event.tags) {
+    tagTitles = event.tags.map((item) => {
+      return item.tag.title;
+    });
+  }
+
+  const icsEvent = {
+    start: [
+      event.startTime.getFullYear(),
+      event.startTime.getMonth(),
+      event.startTime.getDate(),
+      event.startTime.getHours(),
+      event.startTime.getMinutes(),
+    ] as DateArray,
+    end: [
+      event.endTime.getFullYear(),
+      event.endTime.getMonth(),
+      event.endTime.getDate(),
+      event.endTime.getHours(),
+      event.endTime.getMinutes(),
+    ] as DateArray,
+    title: event.name,
+    description: event.description || undefined,
+    location: location.join(", "),
+    url: absoluteEventUrl,
+    // TODO:
+    // status: "CONFIRMED" || "CANCELLED" || "TENTATIVE"
+    // organizer: { name: "", email: "", dir: "any url (Maybe the community profile)"}
+    categories: tagTitles,
+    uid: event.id + event.slug,
+    created: [
+      event.createdAt.getFullYear(),
+      event.createdAt.getMonth(),
+      event.createdAt.getDate(),
+      event.createdAt.getHours(),
+      event.createdAt.getMinutes(),
+    ] as DateArray,
+    lastModified: [
+      event.updatedAt.getFullYear(),
+      event.updatedAt.getMonth(),
+      event.updatedAt.getDate(),
+      event.updatedAt.getHours(),
+      event.updatedAt.getMinutes(),
+    ] as DateArray,
+  };
+
+  const result: unknown = ics.createEvent(icsEvent, (error, icsString) => {
+    if (error) {
+      console.log(error);
+      return null;
+    }
+    return icsString;
+  });
+  return result as string | null;
+}
+
+type LoaderData = Response;
+
+export const loader: LoaderFunction = async (args): Promise<LoaderData> => {
+  const { request, params } = args;
+
+  const slug = getParamValueOrThrow(params, "slug");
+  const event = await getEventBySlugOrThrow(slug);
+  const urlEndingToRemove = "icsDownload";
+  const urlEndingToAppend = "";
+  const absoluteEventURL = transformAbsoluteURL(
+    request.url,
+    urlEndingToRemove,
+    urlEndingToAppend
+  );
+  const ics = createIcsString(event, absoluteEventURL);
+
+  console.log("ICS IN CREATE\n\n", ics);
+
+  return new Response(ics, {
+    status: 200,
+    headers: {
+      "Content-Type": "text/calendar",
+    },
+  });
+};
